Report errors to console

Report errors to console

report2console(file = NULL, line_no = NULL, column = NULL,
  context = NULL, error_message = NULL, advice = NULL,
  build_status = NULL, extra_cat_ante = NULL, extra_cat_post = NULL,
  rstudio = FALSE, log_file = NULL, log_file_sep = "|",
  silent = FALSE, halt = getOption("TeXCheckR.halt_on_error", FALSE),
  as_tbl = getOption("TeXCheckR.error_as_tbl", FALSE))

Arguments

file
The file in which the error occurred.

line_no
The line number locating the source of the error.

column
The position on the line to identify the error (usually following the error).

context
The content of the file, to provide context to the error.

error_message
The error message to display beyond the console.

advice
Advice to the user: how should the detected error be resolved in general?

build_status
What should the build status be reported as?

extra_cat_ante
Character vector extra messages (placed before context).

extra_cat_post
Character vector extra messages (placed after context).

rstudio
If available, should the report be allowed to modify the RStudio session (for example, to pop to the location of the error)?

log_file
Optionally, path to a log file on which error_message will be written.

log_file_sep
How should the log file's fields be separated? By default, with a pipe (as tabs are common within error messages).

silent
(logical, default: FALSE) Suppress all output.

halt
Should failures halt via stop or just display a message in the console?

as_tbl
Return a list. Experimental.